refactor(tui): extract motion scheduling
